requirements.txt
flask
gunicorn
pdf2docx
pytesseract
pdf2image
python-docx
pillow